+++ repl/src/dotty/tools/repl/ReplDriver.scala
@@ -173,6 +173,8 @@ class ReplDriver(settings: Array[String],
 
   private[repl] def replShouldStart: Boolean = shouldStart
 
+  private[repl] def replRendering: Rendering = rendering
+
   private[repl] def replRenderingClassLoader: Option[ClassLoader] =
     Option(rendering).flatMap(loaded => Option(loaded.myClassLoader))
